@@ -110,44 +110,44 @@
   },
   "devDependencies": {
     "async": "2.1.4",
-    "autoprefixer": "6.6.1",
+    "autoprefixer": "6.7.2",
     "babel-eslint": "7.1.1",
     "babel-loader": "6.2.10",
-    "babel-plugin-transform-react-remove-prop-types": "0.2.11",
+    "babel-plugin-transform-react-remove-prop-types": "0.3.2",
     "babel-plugin-transform-runtime": "6.15.0",
-    "babel-runtime": "6.20.0",
-    "brotli-webpack-plugin": "0.1.2",
+    "babel-runtime": "6.22.0",
+    "brotli-webpack-plugin": "0.2.0",
     "compression-webpack-plugin": "0.3.2",
     "css-loader": "0.26.1",
     "csswring": "5.1.1",
-    "eslint": "3.14.0",
-    "eslint-config-airbnb": "14.0.0",
-    "eslint-import-resolver-webpack": "0.8.0",
-    "eslint-plugin-compat": "1.0.0",
-    "eslint-plugin-graphql": "0.4.3",
+    "eslint": "3.15.0",
+    "eslint-config-airbnb": "14.1.0",
+    "eslint-import-resolver-webpack": "0.8.1",
+    "eslint-plugin-compat": "1.0.1",
+    "eslint-plugin-graphql": "0.6.0",
     "eslint-plugin-import": "2.2.0",
     "eslint-plugin-jsx-a11y": "3.0.2",
     "eslint-plugin-react": "6.9.0",
     "exports-loader": "0.6.3",
-    "extract-text-webpack-plugin": "2.0.0-beta.5",
-    "file-loader": "0.9.0",
-    "gemini": "4.16.0",
+    "extract-text-webpack-plugin": "2.0.0-rc.3",
+    "file-loader": "0.10.0",
+    "gemini": "4.18.1",
     "gemini-browserstack": "https://github.com/HSLdevcom/gemini-browserstack",
     "imports-loader": "0.7.0",
     "nightwatch": "0.9.12",
-    "node-sass": "3.13.1",
+    "node-sass": "4.5.0",
     "nodemon": "1.11.0",
     "phantomjs-prebuilt": "2.1.14",
     "postcss-loader": "1.2.2",
     "react-addons-perf": "15.4.2",
     "react-hot-loader": "1.3.1",
-    "sass-loader": "4.1.1",
+    "sass-loader": "5.0.1",
     "selenium-webdriver": "3.0.1",
     "stats-webpack-plugin": "0.4.3",
     "style-loader": "0.13.1",
     "url-loader": "0.5.7",
-    "webpack": "2.2.0-rc.7",
-    "webpack-dev-server": "2.2.0-rc.0",
+    "webpack": "2.2.1",
+    "webpack-dev-server": "2.3.0",
     "webpack-md5-hash": "0.0.5"
   }
 }
